Move TreeNodeSchema and TreeNode to core
