+++ packages/dds/tree/src/simple-tree/api/schemaCreationUtilities.ts
@@ -8,8 +8,8 @@ import { UsageError } from "@fluidframework/telemetry-utils/internal";
 import { fail } from "../../util/index.js";
 
 import type { SchemaFactory, ScopedSchemaName } from "./schemaFactory.js";
-import type { NodeFromSchema, NodeKind, TreeNodeSchemaClass } from "../schemaTypes.js";
-import type { TreeNode } from "../types.js";
+import type { NodeFromSchema } from "../schemaTypes.js";
+import type { NodeKind, TreeNodeSchemaClass, TreeNode } from "../core/index.js";
 import type {
 	InsertableObjectFromSchemaRecord,
 	ObjectFromSchemaRecord,
