+++ packages/dds/tree/src/simple-tree/api/schemaFactory.ts
@@ -37,15 +37,18 @@ import {
 	type ImplicitAllowedTypes,
 	type ImplicitFieldSchema,
 	type InsertableTreeNodeFromImplicitAllowedTypes,
-	type NodeKind,
-	type TreeNodeSchema,
-	type TreeNodeSchemaClass,
 	type FieldProps,
 	createFieldSchema,
 	type DefaultProvider,
 	getDefaultProvider,
 } from "../schemaTypes.js";
-import type { WithType } from "../core/index.js";
+import { inPrototypeChain, TreeNodeValid } from "../core/index.js";
+import type {
+	NodeKind,
+	WithType,
+	TreeNodeSchema,
+	TreeNodeSchemaClass,
+} from "../core/index.js";
 import { type TreeArrayNode, arraySchema } from "../arrayNode.js";
 import {
 	type InsertableObjectFromSchemaRecord,
@@ -70,7 +73,6 @@ import type {
 	TreeObjectNodeUnsafe,
 } from "../typesUnsafe.js";
 import { createFieldSchemaUnsafe } from "./schemaFactoryRecursive.js";
-import { inPrototypeChain, TreeNodeValid } from "../types.js";
 /**
  * Gets the leaf domain schema compatible with a given {@link TreeValue}.
  */
