+++ packages/dds/tree/src/simple-tree/core/index.ts
@@ -5,3 +5,24 @@
 
 export { isTreeNode, TreeNodeKernel, getKernel } from "./treeNodeKernel.js";
 export { type WithType, typeNameSymbol } from "./withType.js";
+export {
+	type TreeChangeEvents,
+	TreeNode,
+	TreeNodeValid,
+	type Unhydrated,
+	type InternalTreeNode,
+	type MostDerivedData,
+	inPrototypeChain,
+} from "./types.js";
+export {
+	type TreeNodeSchema,
+	NodeKind,
+	type TreeNodeSchemaClass,
+	type TreeNodeSchemaNonClass,
+	type TreeNodeSchemaCore,
+} from "./treeNodeSchema.js";
+export {
+	getSimpleNodeSchema,
+	setFlexSchemaFromClassSchema,
+	tryGetSimpleNodeSchema,
+} from "./schemaCaching.js";
