+++ packages/dds/tree/src/simple-tree/core/treeNodeKernel.ts
@@ -5,7 +5,7 @@
 
 import { assert } from "@fluidframework/core-utils/internal";
 import { createEmitter, type Listenable, type Off } from "../../events/index.js";
-import type { TreeChangeEvents, TreeNode, Unhydrated } from "../types.js";
+import type { TreeChangeEvents, TreeNode, Unhydrated } from "./types.js";
 import type { AnchorNode } from "../../core/index.js";
 import {
 	flexTreeSlot,
